terraform: Pass provider schemas to graph builders and interpolator
We need schemas in order to extract references from and interpolate into
HCL2 bodies, so we need to request the provider schemas early enough for
them to be available during graph construction.

For now we briefly start up throwaway provider instances that we use
only for this purpose, since that minimizes the impact on existing
provider init mechanisms. We may try to reuse these instances later once
HCL2 processing is a primary codepath.